작성자 정보처리, 버튼 노출 제어, jsp include

+++ src/main/webapp/jsp/article/detail.jsp
@@ -4,6 +4,7 @@
     
 <%
 	Map<String, Object> articleMap = (Map<String, Object>) request.getAttribute("articleMap");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
 
 <!DOCTYPE html>
@@ -17,14 +18,17 @@
 		<div><%= articleMap.get("id") %>번 게시물 상세보기</div>
 		<div>작성일 : <%= articleMap.get("regDate") %></div>
 		<div>수정일 : <%= articleMap.get("updateDate") %></div>
+		<div>작성자 : <%= articleMap.get("writerName") %></div>
 		<div>제목 : <%= articleMap.get("title") %></div>
 		<div>내용 : <%= articleMap.get("body") %></div>
 	</div>
 	
 	<div>
 		<div><a href="list">목록</a></div>
-		<div><a href="modify?id=<%= articleMap.get("id") %>">수정</a></div>
-		<div><a href="delete?id=<%= articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a></div>
+		<% if ((int) articleMap.get("memberId") == loginedMemberId) { %>
+			<div><a href="modify?id=<%= articleMap.get("id") %>">수정</a></div>
+			<div><a href="delete?id=<%= articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a></div>
+		<% } %>
 	</div>
 </body>
 </html>
